dreame_tuya_trigger_v2/app/requirements.txt
pulsar-client>=3.0.0
pycryptodome>=3.19.0
requests>=2.28.0
